Move widgets to the ui package since they need API access
