+++ packages/drupal-admin-ui/package.json
@@ -60,7 +60,6 @@
     "test": "yarn test:lint && yarn test:unit && yarn test:functional",
     "test:functional": "yarn nightwatch",
     "test:unit": "CI=true JEST_JUNIT_OUTPUT=reports/jest-junit.xml react-scripts test --testResultsProcessor ./node_modules/jest-junit --env=jsdom",
-    "test:lint": "eslint ./src --max-warnings=0 --format junit -o reports/eslint-junit.xml",
     "eject": "react-scripts eject"
   },
   "engines": {
