+++ packages/drupal-admin-ui/src/components/02_atoms/Widgets/StringTextfield.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import TextField from '@material-ui/core/TextField';
-import WidgetPropTypes from '../../../drupal-admin-ui/src/components/05_pages/NodeForm/WidgetPropTypes';
+import WidgetPropTypes from '../../05_pages/NodeForm/WidgetPropTypes';
 
 const StringTextfield = props => (
   <TextField
